Making a utils file
