+++ src/subcommands/update.go
@@ -6,7 +6,6 @@ import (
 	"os"
 	"path"
 	"path/filepath"
-	"strings"
 
 	pb "accretional.com/semantifly/proto/accretional.com/semantifly/proto"
 	"google.golang.org/protobuf/proto"
@@ -43,24 +42,11 @@ func Update(u UpdateArgs) {
 	if u.UpdateCopy == "true" {
 		if err := updateCopy(&u); err != nil {
 			fmt.Printf("Failed to update the copy of the source file: %v", err)
+			return
 		}
 	}
-}
 
-func readIndex(indexFilePath string, index *pb.Index, ignoreIfNotFound ...bool) error {
-	data, err := os.ReadFile(indexFilePath)
-	if err != nil {
-		if os.IsNotExist(err) && (len(ignoreIfNotFound) > 0 && ignoreIfNotFound[0]) {
-			return nil
-		}
-		return fmt.Errorf("failed to read index file: %w", err)
-	}
-
-	if err := proto.Unmarshal(data, index); err != nil {
-		return fmt.Errorf("failed to marshall index file: %w", err)
-	}
-
-	return nil
+	fmt.Printf("Index %s updated successfully to URI %s\n", u.Name, u.DataURI)
 }
 
 func updateIndex(index *pb.Index, u *UpdateArgs) error {
@@ -135,32 +121,4 @@ func updateCopy(u *UpdateArgs) error {
 	}
 
 	return nil
-}
-
-func writeIndex(indexFilePath string, index *pb.Index) error {
-	updatedData, err := proto.Marshal(index)
-	if err != nil {
-		return fmt.Errorf("failed to marshal updated index: %w", err)
-	}
-	if err := os.WriteFile(indexFilePath, updatedData, 0644); err != nil {
-		return fmt.Errorf("failed to write updated index to file: %w", err)
-	}
-
-	return nil
-}
-
-func parseDataType(str string) (pb.DataType, error) {
-	val, ok := pb.DataType_value[strings.ToUpper(str)]
-	if !ok {
-		return pb.DataType_TEXT, fmt.Errorf("unknown data type: %s", str)
-	}
-	return pb.DataType(val), nil
-}
-
-func parseSourceType(str string) (pb.SourceType, error) {
-	val, ok := pb.SourceType_value[strings.ToUpper(str)]
-	if !ok {
-		return pb.SourceType_LOCAL_FILE, fmt.Errorf("unknown source type: %s", str)
-	}
-	return pb.SourceType(val), nil
 }
